chore(build): fix lint + readme

+++ internal/service/service.go
@@ -16,7 +16,7 @@ type UptimeCheckService struct {
 
 func New(provider string, slackToken string, slackChannel string) *UptimeCheckService {
 	var p UptimeProvider
-	switch provider {
+	switch provider { //nolint:gocritic
 	case "mock":
 		p = providers.NewMockUptimeProvider()
 		// TODO add new case(s) for actual uptime monitoring SaaS providers
@@ -26,7 +26,6 @@ func New(provider string, slackToken string, slackChannel string) *UptimeCheckSe
 	if slackToken != "" && slackChannel != "" {
 		slack = NewSlack(slackToken, slackChannel)
 	}
-
 	return &UptimeCheckService{
 		slack:    slack,
 		provider: p,
@@ -36,11 +35,10 @@ func New(provider string, slackToken string, slackChannel string) *UptimeCheckSe
 func (r *UptimeCheckService) Mutate(ctx context.Context, mutation m.Mutation, annotations map[string]string) {
 	check := m.NewUptimeCheck(annotations)
 	if check != nil {
-		switch mutation {
-		case m.CreateOrUpdate:
+		if mutation == m.CreateOrUpdate {
 			err := r.provider.CreateOrUpdateCheck(*check)
 			r.logMutation(ctx, err, mutation, check)
-		case m.Delete:
+		} else if mutation == m.Delete {
 			err := r.provider.DeleteCheck(*check)
 			r.logMutation(ctx, err, mutation, check)
 		}
@@ -51,19 +49,20 @@ func (r *UptimeCheckService) logMutation(ctx context.Context, err error, mutatio
 	if err != nil {
 		msg := fmt.Sprintf("%s of uptime check '%s' (id: %s) failed", string(mutation), check.Name, check.ID)
 		log.FromContext(ctx).Error(err, msg, "check", check)
-		if r.slack != nil {
-			r.slack.SendSlackMessage(ctx, ":large_red_square: "+msg)
+		if r.slack == nil {
+			return
 		}
-
+		r.slack.SendSlackMessage(ctx, ":large_red_square: "+msg)
 	} else {
 		msg := fmt.Sprintf("%s of uptime check '%s' (id: %s) succeeded", string(mutation), check.Name, check.ID)
 		log.FromContext(ctx).Info(msg)
-		if r.slack != nil {
-			emoji := ":large_green_square: "
-			if mutation == m.Delete {
-				emoji = ":warning: "
-			}
-			r.slack.SendSlackMessage(ctx, emoji+msg)
+		if r.slack == nil {
+			return
+		}
+		if mutation == m.Delete {
+			r.slack.SendSlackMessage(ctx, ":warning: "+msg+". Beware: a flood of these delete message may indicate Traefik itself is down!")
+		} else {
+			r.slack.SendSlackMessage(ctx, ":large_green_square: "+msg)
 		}
 	}
 }
